# docsclear/swagger.yaml
basePath: /
definitions:
...
  userdto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  userdto.MeResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  userdto.RegisterUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  utils.AppError:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@goshop.dev
    name: API Support
  description: GoShop - E-commerce API with observability and metrics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GoShop API
  version: 1.0.0
paths:
...
  /auth/me:
    get:
      description: Retourne les informations du profil de l'utilisateur authentifié.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userdto.MeResponse'
        "401":
          description: Non autorisé
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Obtenir le profil utilisateur
      tags:
      - Authentication
...
  /login:
    post:
      consumes:
      - application/json
      description: Authentifie l'utilisateur et retourne les tokens JWT (access +
        refresh).
      parameters:
      - description: Identifiants de connexion
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userdto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens JWT et rôle
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Payload invalide
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Identifiants invalides
          schema:
            $ref: '#/definitions/utils.AppError'
        "429":
          description: Trop de tentatives (Rate Limit)
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Connexion utilisateur
      tags:
      - Authentication
  /logout:
    post:
      description: Révoque la session courante et invalide le token d'actualisation.
      produces:
      - application/json
      responses:
        "200":
          description: '{''message'': ''logged out successfully''}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Non autorisé
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Déconnexion utilisateur
      tags:
      - Authentication
  /register:
    post:
      consumes:
      - application/json
      description: Crée un nouveau compte utilisateur avec un rôle par défaut "user".
      parameters:
      - description: Données d'inscription (email, password)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userdto.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: '{''message'': ''user registered'', ''user_id'': ''uuid''}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Payload invalide
          schema:
            $ref: '#/definitions/utils.AppError'
        "409":
          description: Email déjà enregistré
          schema:
            $ref: '#/definitions/utils.AppError'
        "422":
          description: Validation échouée
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Inscription d'un nouvel utilisateur
      tags:
      - Authentication
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"